feat: add go-parser service (dev)
