+++ docker-compose.dev.yml
@@ -1,5 +1,20 @@
 version: '3'
 services:
+  go-parser:
+    container_name: hdx-oss-dev-go-parser
+    build:
+      context: .
+      dockerfile: ./packages/go-parser/Dockerfile
+    environment:
+      HYPERDX_API_KEY: ${HYPERDX_API_KEY}
+      HYPERDX_LOG_LEVEL: ${HYPERDX_LOG_LEVEL}
+      OTEL_EXPORTER_OTLP_ENDPOINT: http://otel-collector:4318
+      OTEL_LOG_LEVEL: ${HYPERDX_LOG_LEVEL}
+      OTEL_SERVICE_NAME: hdx-oss-dev-go-parser
+    ports:
+      - 7777:7777
+    networks:
+      - internal
   miner:
     container_name: hdx-oss-dev-miner
     build:
