--- packages/api/src/gobin/src/go.mod
@@ -1,5 +1,0 @@
-module sqlObfuscator
-
-go 1.21.5
-
-require github.com/DataDog/go-sqllexer v0.0.10
